requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
billiard==3.6.4.0
blinker==1.5
Brotli==1.0.9
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.0
Django==4.1.5
greenlet==2.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
kaitaistruct==0.10
kombu==5.2.4
loguru==0.6.0
outcome==1.2.0
packaging==23.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
pyOpenSSL==23.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.21.0
python-http-client==3.3.7
pytz==2022.7.1
requests==2.28.2
selenium==4.7.2
sendgrid==6.9.7
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.46
SQLAlchemy-Utils==0.39.0
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
webdriver-manager==3.8.5
wsproto==1.2.0
zstandard==0.19.0
